+++ rust/crd/src/lib.rs
@@ -132,6 +132,12 @@ impl FlaskAppConfigOptions for AirflowConfigOptions {
     }
 }
 
+/// An Airflow cluster stacklet. This resource is managed by the Stackable operator for Apache Airflow.
+/// Find more information on how to use it and the resources that the operator generates in the
+/// [operator documentation](DOCS_BASE_URL_PLACEHOLDER/airflow/).
+///
+/// The CRD contains three roles: webserver, scheduler and worker/celeryExecutor.
+/// You can use either the celeryExecutor or the kubernetesExecutor.
 #[derive(Clone, CustomResource, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[kube(
     group = "airflow.stackable.tech",
@@ -149,19 +155,23 @@ impl FlaskAppConfigOptions for AirflowConfigOptions {
 )]
 #[serde(rename_all = "camelCase")]
 pub struct AirflowClusterSpec {
-    /// The Airflow image to use
+    // no doc string - See ProductImage struct
     pub image: ProductImage,
 
-    /// Global cluster configuration that applies to all roles and role groups
+    /// Configuration that applies to all roles and role groups.
+    /// This includes settings for authentication, git sync, service exposition and volumes, among other things.
     pub cluster_config: AirflowClusterConfig,
 
-    /// Cluster operations like pause reconciliation or cluster stop.
+    // no doc string - See ClusterOperation struct
     #[serde(default)]
     pub cluster_operation: ClusterOperation,
 
+    /// The `webserver` role provides the main UI for user interaction.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub webservers: Option<Role<AirflowConfigFragment>>,
 
+    /// The `scheduler` is responsible for triggering jobs and persisting their metadata to the backend database.
+    /// Jobs are scheduled on the workers/executors.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub schedulers: Option<Role<AirflowConfigFragment>>,
 
@@ -174,18 +184,23 @@ pub struct AirflowClusterSpec {
 pub struct AirflowClusterConfig {
     #[serde(flatten)]
     pub authentication: AirflowAuthentication,
-
+    /// The name of the Secret object containing the admin user credentials and database connection details.
+    /// Read the
+    /// [getting started guide first steps](DOCS_BASE_URL_PLACEHOLDER/airflow/getting_started/first_steps)
+    /// to find out more.
     pub credentials_secret: String,
-
+    /// The `gitSync` settings allow configuring DAGs to mount via `git-sync`.
+    /// Learn more in the
+    /// [mounting DAGs documentation](DOCS_BASE_URL_PLACEHOLDER/airflow/usage-guide/mounting-dags#_via_git_sync).
     #[serde(default)]
     pub dags_git_sync: Vec<GitSync>,
-
+    /// for internal use only - not for production use.
     #[serde(default)]
     pub expose_config: bool,
-
+    /// Whether to load example DAGs or not; defaults to false. The examples are used in the
+    /// [getting started guide](DOCS_BASE_URL_PLACEHOLDER/airflow/getting_started/).
     #[serde(default)]
     pub load_examples: bool,
-
     /// This field controls which type of Service the Operator creates for this AirflowCluster:
     ///
     /// * cluster-internal: Use a ClusterIP service
@@ -195,19 +210,23 @@ pub struct AirflowClusterConfig {
     /// * external-stable: Use a LoadBalancer service
     ///
     /// This is a temporary solution with the goal to keep yaml manifests forward compatible.
-    /// In the future, this setting will control which ListenerClass <https://docs.stackable.tech/home/stable/listener-operator/listenerclass.html>
+    /// In the future, this setting will control which [ListenerClass](DOCS_BASE_URL_PLACEHOLDER/listener-operator/listenerclass.html)
     /// will be used to expose the service, and ListenerClass names will stay the same, allowing for a non-breaking change.
     #[serde(default)]
     pub listener_class: CurrentlySupportedListenerClasses,
 
-    /// Name of the Vector aggregator discovery ConfigMap.
+    /// Name of the Vector aggregator [discovery ConfigMap](DOCS_BASE_URL_PLACEHOLDER/concepts/service_discovery).
     /// It must contain the key `ADDRESS` with the address of the Vector aggregator.
+    /// Follow the [logging tutorial](DOCS_BASE_URL_PLACEHOLDER/tutorials/logging-vector-aggregator)
+    /// to learn how to configure log aggregation with Vector.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub vector_aggregator_config_map_name: Option<String>,
 
+    /// Additional volumes to define. Use together with `volumeMounts` to mount the volumes.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub volumes: Option<Vec<Volume>>,
 
+    /// Additional volumes to mount. Use together with `volumes` to define volumes.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub volume_mounts: Option<Vec<VolumeMount>>,
 }
@@ -352,12 +371,15 @@ impl AirflowRole {
 
 #[derive(Clone, Debug, Deserialize, Display, JsonSchema, PartialEq, Serialize)]
 pub enum AirflowExecutor {
+    /// The celery executor.
+    /// Deployed with an explicit number of replicas.
     #[serde(rename = "celeryExecutors")]
     CeleryExecutor {
         #[serde(flatten)]
         config: Role<AirflowConfigFragment>,
     },
 
+    /// With the Kuberentes executor, executor Pods are created on demand.
     #[serde(rename = "kubernetesExecutors")]
     KubernetesExecutor {
         #[serde(flatten)]