+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- BREAKING: Fixed various issues in the CRD structure. `clusterConfig.credentialsSecret` is now mandatory ([#353]).
+
+[#353]: https://github.com/stackabletech/airflow-operator/pull/353
+
 ## [23.11.0] - 2023-11-24
 
 ### Added