+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Increase resource defaults ([#352]).
+
 ### Fixed
 
 - BREAKING: Fixed various issues in the CRD structure. `clusterConfig.credentialsSecret` is now mandatory ([#353]).
 
+[#352]: https://github.com/stackabletech/airflow-operator/pull/352
 [#353]: https://github.com/stackabletech/airflow-operator/pull/353
 
 ## [23.11.0] - 2023-11-24