+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - `operator-rs` `0.56.1` -> `0.57.0` ([#354]).
+- Increase resource defaults ([#352]).
 
 ### Fixed
 
@@ -16,6 +17,7 @@
 
 [#354]: https://github.com/stackabletech/airflow-operator/pull/354
 [#353]: https://github.com/stackabletech/airflow-operator/pull/353
+[#352]: https://github.com/stackabletech/airflow-operator/pull/352
 
 ## [23.11.0] - 2023-11-24