+++ CHANGELOG.md
@@ -15,9 +15,9 @@
 
 - BREAKING: Fixed various issues in the CRD structure. `clusterConfig.credentialsSecret` is now mandatory ([#353]).
 
-[#354]: https://github.com/stackabletech/airflow-operator/pull/354
-[#353]: https://github.com/stackabletech/airflow-operator/pull/353
 [#352]: https://github.com/stackabletech/airflow-operator/pull/352
+[#353]: https://github.com/stackabletech/airflow-operator/pull/353
+[#354]: https://github.com/stackabletech/airflow-operator/pull/354
 
 ## [23.11.0] - 2023-11-24